requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pywin32; platform_system=="Windows"
pypdf==4.3.1
pystray==0.19.5
Pillow==12.0.0
pyperclip==1.11.0
redis==5.2.1
arq==0.26.3
sentry-sdk[fastapi]==2.18.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
